+++ crates/fuel-core/src/graphql_api/database.rs
@@ -182,26 +182,7 @@ impl DatabaseMessageProof for ReadView {
     }
 }
 
-impl OnChainDatabase for ReadView {
-    fn owned_message_ids(
-        &self,
-        owner: &Address,
-        start_message_id: Option<Nonce>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<Nonce>> {
-        self.on_chain
-            .owned_message_ids(owner, start_message_id, direction)
-    }
-
-    fn owned_coins_ids(
-        &self,
-        owner: &Address,
-        start_coin: Option<UtxoId>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<UtxoId>> {
-        self.on_chain.owned_coins_ids(owner, start_coin, direction)
-    }
-}
+impl OnChainDatabase for ReadView {}
 
 impl OffChainDatabase for ReadView {
     fn block_height(&self, block_id: &BlockId) -> StorageResult<BlockHeight> {
@@ -210,6 +191,25 @@ impl OffChainDatabase for ReadView {
 
     fn tx_status(&self, tx_id: &TxId) -> StorageResult<TransactionStatus> {
         self.off_chain.tx_status(tx_id)
+    }
+
+    fn owned_coins_ids(
+        &self,
+        owner: &Address,
+        start_coin: Option<UtxoId>,
+        direction: IterDirection,
+    ) -> BoxedIter<'_, StorageResult<UtxoId>> {
+        self.off_chain.owned_coins_ids(owner, start_coin, direction)
+    }
+
+    fn owned_message_ids(
+        &self,
+        owner: &Address,
+        start_message_id: Option<Nonce>,
+        direction: IterDirection,
+    ) -> BoxedIter<'_, StorageResult<Nonce>> {
+        self.off_chain
+            .owned_message_ids(owner, start_message_id, direction)
     }
 
     fn owned_transactions_ids(