+++ crates/fuel-core/src/graphql_api/storage.rs
@@ -1,6 +1,8 @@
 use fuel_core_storage::kv_store::StorageColumn;
 
 pub mod blocks;
+pub mod coins;
+pub mod messages;
 pub mod transactions;
 
 /// GraphQL database tables column ids to the corresponding [`fuel_core_storage::Mappable`] table.