+++ crates/fuel-core/src/p2p_test_helpers.rs
@@ -5,7 +5,6 @@ use crate::{
     database::Database,
     p2p::Multiaddr,
     service::{
-        genesis::execute_and_commit_genesis_block,
         Config,
         FuelService,
         ServiceTrait,
@@ -399,7 +398,7 @@ pub async fn make_node(node_config: Config, test_txs: Vec<Transaction>) -> Node 
 async fn extract_p2p_config(node_config: &Config) -> fuel_core_p2p::config::Config {
     let bootstrap_config = node_config.p2p.clone();
     let db = Database::in_memory();
-    execute_and_commit_genesis_block(node_config, &db)
+    crate::service::genesis::execute_and_commit_genesis_block(node_config, &db)
         .await
         .unwrap();
     bootstrap_config