+++ crates/fuel-core/src/service/adapters/graphql_api/on_chain.rs
@@ -27,11 +27,7 @@ use fuel_core_types::{
         primitives::DaBlockHeight,
     },
     entities::message::Message,
-    fuel_tx::{
-        Address,
-        AssetId,
-        UtxoId,
-    },
+    fuel_tx::AssetId,
     fuel_types::{
         BlockHeight,
         Nonce,
@@ -106,26 +102,4 @@ impl DatabaseChain for Database {
     }
 }
 
-impl OnChainDatabase for Database {
-    fn owned_message_ids(
-        &self,
-        owner: &Address,
-        start_message_id: Option<Nonce>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<Nonce>> {
-        self.owned_message_ids(owner, start_message_id, Some(direction))
-            .map(|result| result.map_err(StorageError::from))
-            .into_boxed()
-    }
-
-    fn owned_coins_ids(
-        &self,
-        owner: &Address,
-        start_coin: Option<UtxoId>,
-        direction: IterDirection,
-    ) -> BoxedIter<'_, StorageResult<UtxoId>> {
-        self.owned_coins_ids(owner, start_coin, Some(direction))
-            .map(|res| res.map_err(StorageError::from))
-            .into_boxed()
-    }
-}
+impl OnChainDatabase for Database {}