+++ crates/storage/src/column.rs
@@ -63,13 +63,7 @@ pub enum Column {
     /// The column id of metadata about the blockchain
     Metadata = 17,
     /// See [`SealedBlockConsensus`](crate::tables::SealedBlockConsensus)
-    FuelBlockConsensus = 18,
-
-    // Below are not required tables. They are used for API and may be removed or moved to another place in the future.
-    /// The column of the table that stores `true` if `owner` owns `Coin` with `coin_id`
-    OwnedCoins = 19,
-    /// The column of the table that stores `true` if `owner` owns `Message` with `message_id`
-    OwnedMessageIds = 20,
+    FuelBlockConsensus = 19,
 }
 
 impl Column {